Task Finished
Task List Finished
From quote list, select quote, then tap “back”.  The GetQuoteList api should be called one time.  The log shows it is being called 3 times.

In the quote, if you select a priority it should call update quote.  It doesn’t

Get Quote returns a value for Priority. This value should be displayed as the default Prioirty. Right now it just says “Select an Item”.  If the value is “normal”, it should say “normal”

Expense lable and number is red. It should be black like everything else.

The amount provided by getQuote is not being displayed.  Instead there is only a dollar sign.  The app should display exactly what is provided in the Amount field.  Do not add any dollar sign or anything.